refactor: Extract a file module for file-based paths
